requirements.txt
anthropic>=0.25.0
elevenlabs>=1.0.0
openai-whisper
pvporcupine
pyaudio
sounddevice
soundfile
numpy
PyQt5
pyautogui
python-dotenv
pygame
pycaw
comtypes
